fix(AscensionSummary): delete old fix(AscensionSummary): delete old and Material
